Внесение изменений в весь проект

+++ src/main/java/web/controller/UserController.java
@@ -1,56 +1,66 @@
 package web.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.ModelMap;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import web.model.User;
 import web.service.UserService;
 
+import javax.validation.Valid;
 
 @Controller
 public class UserController {
-    private UserService userService;
 
-    @Autowired
+    private final UserService userService;
+
     public UserController(UserService userService) {
         this.userService = userService;
     }
 
     @GetMapping("/")
-    public String printUsers(ModelMap model) {
-        model.addAttribute("list", userService.getAllUsers());
+    public String getAllUsers(Model model) {
+        model.addAttribute("messages", userService.getAllUsers());
         return "users";
     }
 
     @GetMapping("/add")
-    public String showAddForm(ModelMap model) {
-        model.addAttribute("user", new User());
-        return "user_form";
+    public String CreateUserForm(@ModelAttribute("user") User user) {
+        return "adduser";
     }
 
-    @PostMapping("/add")
-    public String addUser(@ModelAttribute User user) {
+    @PostMapping("/adduser")
+    public String addUser(@ModelAttribute("user") @Valid User user, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "adduser";
+        }
         userService.addUser(user);
-        return "redirect:/users";
+        return "redirect:/";
     }
 
     @GetMapping("/edit/{id}")
-    public String showEditForm(@PathVariable("id") Long id, ModelMap model) {
+    public String getUser(@PathVariable("id") Long id, Model model) {
         User user = userService.getUserById(id);
         model.addAttribute("user", user);
-        return "user_form";
+        return "updateuser";
     }
 
-    @PostMapping("/edit")
-    public String editUser(@ModelAttribute User user) {
+    @PostMapping("/updateuser")
+    public String updateUser(@ModelAttribute("user") @Valid User user, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "updateuser";
+        }
         userService.updateUser(user);
-        return "redirect:/users";
+        return "redirect:/";
     }
 
-    @GetMapping("/delete/{id}")
-    public String deleteUser(@PathVariable("id") Long id) {
+    // РАБОТАЕТ!
+    @GetMapping("/deleteUser")
+    public String deleteUser(@ModelAttribute("id") @Valid Long id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "deleteuser";
+        }
         userService.deleteUserById(id);
-        return "redirect:/users";
+        return "redirect:/";
     }
 }
